+++ tests/testmodels.py
@@ -664,3 +664,8 @@ class Principal(Model):
 
 class Signals(Model):
     name = fields.CharField(max_length=255)
+
+
+class DefaultUpdate(Model):
+    created_at = fields.DatetimeField(auto_now_add=True)
+    updated_at = fields.DatetimeField(auto_now=True)